refactor : SearchControls 하나의 컴포넌트로 분리
