+++ src/SearchDialog.tsx
@@ -1,25 +1,15 @@
-import { useCallback, useEffect, useMemo, useRef, useState, memo } from "react";
+import { useEffect, useMemo, useRef, useState, memo, useCallback } from "react";
 import {
   Box,
   Button,
-  Checkbox,
-  CheckboxGroup,
-  FormControl,
-  FormLabel,
   HStack,
-  Input,
   Modal,
   ModalBody,
   ModalCloseButton,
   ModalContent,
   ModalHeader,
   ModalOverlay,
-  Select,
-  Stack,
   Table,
-  Tag,
-  TagCloseButton,
-  TagLabel,
   Tbody,
   Td,
   Text,
@@ -27,13 +17,13 @@ import {
   Thead,
   Tr,
   VStack,
-  Wrap,
 } from "@chakra-ui/react";
 import { useScheduleContext } from "./ScheduleContext.tsx";
 import { Lecture } from "./types.ts";
 import { parseSchedule } from "./utils.ts";
 import axios from "axios";
-import { DAY_LABELS } from "./constants.ts";
+import { useAutoCallback } from "./hooks/useAutoCallback.ts";
+import SearchControls from "./components/SearchFilters/SearchControls.tsx";
 
 interface Props {
   searchInfo: {
@@ -44,7 +34,7 @@ interface Props {
   onClose: () => void;
 }
 
-interface SearchOption {
+export interface SearchOption {
   query?: string;
   grades: number[];
   days: string[];
@@ -52,33 +42,6 @@ interface SearchOption {
   majors: string[];
   credits?: number;
 }
-
-const TIME_SLOTS = [
-  { id: 1, label: "09:00~09:30" },
-  { id: 2, label: "09:30~10:00" },
-  { id: 3, label: "10:00~10:30" },
-  { id: 4, label: "10:30~11:00" },
-  { id: 5, label: "11:00~11:30" },
-  { id: 6, label: "11:30~12:00" },
-  { id: 7, label: "12:00~12:30" },
-  { id: 8, label: "12:30~13:00" },
-  { id: 9, label: "13:00~13:30" },
-  { id: 10, label: "13:30~14:00" },
-  { id: 11, label: "14:00~14:30" },
-  { id: 12, label: "14:30~15:00" },
-  { id: 13, label: "15:00~15:30" },
-  { id: 14, label: "15:30~16:00" },
-  { id: 15, label: "16:00~16:30" },
-  { id: 16, label: "16:30~17:00" },
-  { id: 17, label: "17:00~17:30" },
-  { id: 18, label: "17:30~18:00" },
-  { id: 19, label: "18:00~18:50" },
-  { id: 20, label: "18:55~19:45" },
-  { id: 21, label: "19:50~20:40" },
-  { id: 22, label: "20:45~21:35" },
-  { id: 23, label: "21:40~22:30" },
-  { id: 24, label: "22:35~23:25" },
-];
 
 const PAGE_SIZE = 100;
 
@@ -191,25 +154,6 @@ const filterLectures = (
   });
 };
 
-// 체크박스 그룹 컴포넌트를 메모이제이션
-const MemoizedCheckboxGroup = memo(
-  ({
-    children,
-    value,
-    onChange,
-    colorScheme,
-  }: {
-    children: React.ReactNode;
-    value: any[];
-    onChange: (value: any[]) => void;
-    colorScheme?: string;
-  }) => (
-    <CheckboxGroup value={value} onChange={onChange} colorScheme={colorScheme}>
-      {children}
-    </CheckboxGroup>
-  )
-);
-
 const SearchDialog = memo(({ searchInfo, onClose }: Props) => {
   const { setSchedulesMap } = useScheduleContext();
 
@@ -249,41 +193,66 @@ const SearchDialog = memo(({ searchInfo, onClose }: Props) => {
   }, [cachedLectures]);
 
   // 검색 옵션 변경 핸들러 메모이제이션
-  const changeSearchOption = useCallback(
+  const changeSearchOption = useAutoCallback(
     (field: keyof SearchOption, value: SearchOption[typeof field]) => {
       setPage(1);
       setSearchOptions((prev) => ({ ...prev, [field]: value }));
       loaderWrapperRef.current?.scrollTo(0, 0);
-    },
-    []
+    }
   );
 
-  const addSchedule = useCallback(
-    (lecture: CachedLecture) => {
-      if (!searchInfo) return;
-
-      const { tableId } = searchInfo;
-      const schedules = lecture.parsedSchedules.map((schedule) => ({
-        ...schedule,
-        lecture: {
-          id: lecture.id,
-          title: lecture.title,
-          grade: lecture.grade,
-          credits: lecture.credits,
-          major: lecture.major,
-          schedule: lecture.schedule,
-        },
-      }));
-
-      setSchedulesMap((prev) => ({
-        ...prev,
-        [tableId]: [...prev[tableId], ...schedules],
-      }));
-
-      onClose();
+  // 각 필터별 개별 핸들러 함수들
+  const handleGradeChange = useCallback(
+    (value: number[]) => {
+      changeSearchOption("grades", value);
     },
-    [searchInfo, setSchedulesMap, onClose]
+    [changeSearchOption]
   );
+
+  const handleDayChange = useCallback(
+    (value: string[]) => {
+      changeSearchOption("days", value);
+    },
+    [changeSearchOption]
+  );
+
+  const handleTimeChange = useCallback(
+    (values: number[]) => {
+      changeSearchOption("times", values);
+    },
+    [changeSearchOption]
+  );
+
+  const handleMajorChange = useCallback(
+    (values: string[]) => {
+      changeSearchOption("majors", values);
+    },
+    [changeSearchOption]
+  );
+
+  const addSchedule = useAutoCallback((lecture: CachedLecture) => {
+    if (!searchInfo) return;
+
+    const { tableId } = searchInfo;
+    const schedules = lecture.parsedSchedules.map((schedule) => ({
+      ...schedule,
+      lecture: {
+        id: lecture.id,
+        title: lecture.title,
+        grade: lecture.grade,
+        credits: lecture.credits,
+        major: lecture.major,
+        schedule: lecture.schedule,
+      },
+    }));
+
+    setSchedulesMap((prev) => ({
+      ...prev,
+      [tableId]: [...prev[tableId], ...schedules],
+    }));
+
+    onClose();
+  });
 
   // API 호출 최적화
   useEffect(() => {
@@ -339,171 +308,34 @@ const SearchDialog = memo(({ searchInfo, onClose }: Props) => {
         <ModalCloseButton />
         <ModalBody>
           <VStack spacing={4} align="stretch">
-            <HStack spacing={4}>
-              <FormControl>
-                <FormLabel>검색어</FormLabel>
-                <Input
-                  placeholder="과목명 또는 과목코드"
-                  value={searchOptions.query}
-                  onChange={(e) => changeSearchOption("query", e.target.value)}
-                />
-              </FormControl>
+            <SearchControls.SearchInputFilter
+              searchOptions={searchOptions}
+              onChange={changeSearchOption}
+            />
 
-              <FormControl>
-                <FormLabel>학점</FormLabel>
-                <Select
-                  value={searchOptions.credits}
-                  onChange={(e) =>
-                    changeSearchOption("credits", e.target.value)
-                  }
-                >
-                  <option value="">전체</option>
-                  <option value="1">1학점</option>
-                  <option value="2">2학점</option>
-                  <option value="3">3학점</option>
-                </Select>
-              </FormControl>
+            <HStack spacing={4}>
+              <SearchControls.GradeFilter
+                value={searchOptions.grades}
+                onChange={handleGradeChange}
+              />
+              <SearchControls.DayFilter
+                value={searchOptions.days}
+                onChange={handleDayChange}
+              />
             </HStack>
 
             <HStack spacing={4}>
-              <FormControl>
-                <FormLabel>학년</FormLabel>
-                <MemoizedCheckboxGroup
-                  value={searchOptions.grades}
-                  onChange={(value) =>
-                    changeSearchOption("grades", value.map(Number))
-                  }
-                >
-                  <HStack spacing={4}>
-                    {[1, 2, 3, 4].map((grade) => (
-                      <Checkbox key={grade} value={grade}>
-                        {grade}학년
-                      </Checkbox>
-                    ))}
-                  </HStack>
-                </MemoizedCheckboxGroup>
-              </FormControl>
-
-              <FormControl>
-                <FormLabel>요일</FormLabel>
-                <MemoizedCheckboxGroup
-                  value={searchOptions.days}
-                  onChange={(value) =>
-                    changeSearchOption("days", value as string[])
-                  }
-                >
-                  <HStack spacing={4}>
-                    {DAY_LABELS.map((day) => (
-                      <Checkbox key={day} value={day}>
-                        {day}
-                      </Checkbox>
-                    ))}
-                  </HStack>
-                </MemoizedCheckboxGroup>
-              </FormControl>
+              <SearchControls.TimeFilter
+                value={searchOptions.times}
+                onChange={handleTimeChange}
+              />
+              <SearchControls.MajorFilter
+                value={searchOptions.majors}
+                onChange={handleMajorChange}
+                allMajors={allMajors}
+              />
             </HStack>
 
-            <HStack spacing={4}>
-              <FormControl>
-                <FormLabel>시간</FormLabel>
-                <MemoizedCheckboxGroup
-                  colorScheme="green"
-                  value={searchOptions.times}
-                  onChange={(values) =>
-                    changeSearchOption("times", values.map(Number))
-                  }
-                >
-                  <Wrap spacing={1} mb={2}>
-                    {searchOptions.times
-                      .sort((a, b) => a - b)
-                      .map((time) => (
-                        <Tag
-                          key={time}
-                          size="sm"
-                          variant="outline"
-                          colorScheme="blue"
-                        >
-                          <TagLabel>{time}교시</TagLabel>
-                          <TagCloseButton
-                            onClick={() =>
-                              changeSearchOption(
-                                "times",
-                                searchOptions.times.filter((v) => v !== time)
-                              )
-                            }
-                          />
-                        </Tag>
-                      ))}
-                  </Wrap>
-                  <Stack
-                    spacing={2}
-                    overflowY="auto"
-                    h="100px"
-                    border="1px solid"
-                    borderColor="gray.200"
-                    borderRadius={5}
-                    p={2}
-                  >
-                    {TIME_SLOTS.map(({ id, label }) => (
-                      <Box key={id}>
-                        <Checkbox key={id} size="sm" value={id}>
-                          {id}교시({label})
-                        </Checkbox>
-                      </Box>
-                    ))}
-                  </Stack>
-                </MemoizedCheckboxGroup>
-              </FormControl>
-
-              <FormControl>
-                <FormLabel>전공</FormLabel>
-                <MemoizedCheckboxGroup
-                  colorScheme="green"
-                  value={searchOptions.majors}
-                  onChange={(values) =>
-                    changeSearchOption("majors", values as string[])
-                  }
-                >
-                  <Wrap spacing={1} mb={2}>
-                    {searchOptions.majors.map((major) => (
-                      <Tag
-                        key={major}
-                        size="sm"
-                        variant="outline"
-                        colorScheme="blue"
-                      >
-                        <TagLabel>{major.split("<p>").pop()}</TagLabel>
-                        <TagCloseButton
-                          onClick={() =>
-                            changeSearchOption(
-                              "majors",
-                              searchOptions.majors.filter((v) => v !== major)
-                            )
-                          }
-                        />
-                      </Tag>
-                    ))}
-                  </Wrap>
-                  <Stack
-                    spacing={2}
-                    overflowY="auto"
-                    h="100px"
-                    border="1px solid"
-                    borderColor="gray.200"
-                    borderRadius={5}
-                    p={2}
-                  >
-                    {allMajors.map((major) => (
-                      <Box key={major}>
-                        <Checkbox key={major} size="sm" value={major}>
-                          {major.replace(/<p>/gi, " ")}
-                        </Checkbox>
-                      </Box>
-                    ))}
-                  </Stack>
-                </MemoizedCheckboxGroup>
-              </FormControl>
-            </HStack>
             <Text align="right">검색결과: {filteredLectures.length}개</Text>
             <Box>
               <Table>
